refactor: move ChangelogButton to header/ directory
ChangelogButton is a header component, not an import component.
